+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Change Log
 Version 1.1.0 *(2019-UNKNOWN)*
 -----------------------------
 * New: Add PopKornCompat to use PopKorn from java code
-* New: support `internal` modifier for Injectable classes
+* New: Supports `internal` modifier for Injectable classes
+* New: Supports incremental annotation processing
     
 
 Version 1.0.1 *(2019-11-20)*